// .PRD/erp-refactor/v3/types/bom-update-tool.ts
import { z } from 'zod';

export const BomUpdateToolSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  current_bom: z.string(),
  new_bom: z.string(),
});

export type BomUpdateTool = z.infer<typeof BomUpdateToolSchema>;

export const BomUpdateToolInsertSchema = BomUpdateToolSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type BomUpdateToolInsert = z.infer<typeof BomUpdateToolInsertSchema>;